# 1.api-gateway/src/docs/openapi/v1/combined.yaml
openapi: 3.0.3
info:
  title: eServices API Gateway (combined)
  version: 1.0.0
  description: Aggregated API for all microservices
servers:
  - url: /
    description: API Gateway
paths:
...
  /v1/auth/verify-email:
    post:
      summary: Verify email
      operationId: Verifyemail
      tags:
        - Auth
        - Local
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Verify email
              required:
                - code
              properties:
                code:
                  type: string
                  example: '123456'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                description: Verify email response
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Email verified successfully
        '400':
          description: Invalid or expired verification code
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Invalid or expired verification code
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Server error
      security: []
...
  /v1/audit-logs:
    get:
      tags:
        - Audit
      summary: List audit logs
      description: Supports pagination, sorting, and simple filtering.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: sort
          schema:
            type: string
            example: '-createdAt'
          description: Sort by field (prepend '-' for descending)
        - in: query
          name: service
          schema:
            type: string
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 2
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: 66b5f1c0f7d5a1f2e1a3b456
                        service:
                          type: string
                          example: api-gateway
                        action:
                          type: string
                          example: user.deactivate
                        actor:
                          type: object
                          properties:
                            id:
                              type: string
                              example: 64fa0a8f…
                            type:
                              type: string
                              example: user
                        resource:
                          type: object
                          properties:
                            type:
                              type: string
                              example: user
                            id:
                              type: string
                              example: 64fa0a8f…
                        ip:
                          type: string
                          example: 203.0.113.42
                        userAgent:
                          type: string
                          example: Mozilla/5.0 (...)
                        meta:
                          type: object
                          additionalProperties: true
                        createdAt:
                          type: string
                          format: date-time
                      required:
                        - id
                        - action
                        - createdAt
                required:
                  - count
                  - data
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal server error
    post:
      tags:
        - Audit
      summary: Create an audit log entry
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                service:
                  type: string
                  example: api-gateway
                action:
                  type: string
                  example: user.deactivate
                actor:
                  type: object
                  properties:
                    id:
                      type: string
                      example: 64fa0a8f…
                    type:
                      type: string
                      example: user
                resource:
                  type: object
                  properties:
                    type:
                      type: string
                      example: user
                    id:
                      type: string
                      example: 64fa0a8f…
                ip:
                  type: string
                  example: 203.0.113.42
                userAgent:
                  type: string
                meta:
                  type: object
                  additionalProperties: true
              required:
                - action
      responses:
        '201':
          description: Audit log created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 66b5f1c0f7d5a1f2e1a3b456
                      service:
                        type: string
                        example: api-gateway
                      action:
                        type: string
                        example: user.deactivate
                      actor:
                        type: object
                        properties:
                          id:
                            type: string
                            example: 64fa0a8f…
                          type:
                            type: string
                            example: user
                      resource:
                        type: object
                        properties:
                          type:
                            type: string
                            example: user
                          id:
                            type: string
                            example: 64fa0a8f…
                      ip:
                        type: string
                        example: 203.0.113.42
                      userAgent:
                        type: string
                        example: Mozilla/5.0 (...)
                      meta:
                        type: object
                        additionalProperties: true
                      createdAt:
                        type: string
                        format: date-time
                    required:
                      - id
                      - action
                      - createdAt
                required:
                  - success
                  - data
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal server error
  /v1/audit-logs/{id}:
    get:
      tags:
        - Audit
      summary: Get an audit log entry by ID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Audit log ID
      responses:
        '200':
          description: Single audit log
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 66b5f1c0f7d5a1f2e1a3b456
                      service:
                        type: string
                        example: api-gateway
                      action:
                        type: string
                        example: user.deactivate
                      actor:
                        type: object
                        properties:
                          id:
                            type: string
                            example: 64fa0a8f…
                          type:
                            type: string
                            example: user
                      resource:
                        type: object
                        properties:
                          type:
                            type: string
                            example: user
                          id:
                            type: string
                            example: 64fa0a8f…
                      ip:
                        type: string
                        example: 203.0.113.42
                      userAgent:
                        type: string
                        example: Mozilla/5.0 (...)
                      meta:
                        type: object
                        additionalProperties: true
                      createdAt:
                        type: string
                        format: date-time
                    required:
                      - id
                      - action
                      - createdAt
                required:
                  - success
                  - data
        '401':
          description: Unauthorized
        '404':
          description: Audit log not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Internal server error
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []